requirements.txt
python-dotenv
SQLAlchemy
PyMySQL